app/backend/requirements.txt
asgiref==3.3.0
Django==3.0.5
djangorestframework==3.12.2
djongo==1.3.3
dnspython==2.0.0
pymongo==3.11.0
pytz==2020.4
PyYAML==5.3.1
sqlparse==0.2.4
uritemplate==3.0.1
django-cors-headers==3.5.0
drf-yasg==1.20.0
boto3==1.16.32
django-storages==1.10.1
cryptography==3.3.1
nltk==3.5
python-datamuse==1.3.0
django-unixdatetimefield==1.2.0
